refactor: moved subgraph hooks to common folder
