+++ apps/davi/src/stores/modules/SnapshotERC20Guild/index.ts
@@ -15,7 +15,7 @@ import {
   useGetAllPermissions as useGetAllPermissionsFromSubgraph,
   useGetVotes as useGetVotesFromSubgraph,
   useProposalVotesOfVoter as useProposalVotesOfVoterFromSubgraph,
-} from '../subgraph/common';
+} from '../common/fetchers/subgraph';
 
 import {
   useProposal,
