+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGetAllPermissions/index.ts
