+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGetAllPermissions/query.graphql
