+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGetMemberList/query.graphql
