+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGetNumberOfActiveProposals/query.graphql
