+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGetVotes/index.ts
