+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useGuildConfig/query.graphql
