+++ apps/davi/src/stores/modules/common/fetchers/subgraph/useProposalVotesOfVoter/index.ts
